+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 * Interface to [deepchem](https://deepchem.io/) to generate regression models.
 * Interface to [CAT](https://github.com/nlesc-nano/CAT) to compute solvation energies and activity coefficients.
+* Train or predict values for a set of smiles using a statistical model.